+++ app/src/at-phutran/res/layout/activity_view.xml
@@ -9,7 +9,7 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="110dp"
-        android:background="@drawable/background">
+        android:background="@drawable/bg_">
     </RelativeLayout>
     <LinearLayout
         android:layout_width="match_parent"
@@ -43,7 +43,7 @@
                 android:layout_weight="1"
                 android:layout_width="0dp"
                 android:layout_height="10dp"
-                android:src="@drawable/dot"
+                android:src="@drawable/ic_dot"
                 android:layout_marginRight="10dp"
                 android:layout_marginEnd="10dp"
                 android:contentDescription="@string/todo" />
@@ -83,7 +83,7 @@
                 android:layout_width="0dp"
                 android:layout_height="25dp"
                 android:padding="7dp"
-                android:src="@drawable/dot1"
+                android:src="@drawable/ic_dot1"
                 android:scaleType="centerInside"
                 android:layout_weight="1"
                 android:contentDescription="@string/todo" />
