+++ app/src/main/AndroidManifest.xml
@@ -16,8 +16,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ListActivity" />
-        <activity android:name=".ViewActivity"></activity>
     </application>
 
 </manifest>
