+++ app/src/main/java/asiantech/internship/summer/ViewActivity.java
@@ -1,51 +1,6 @@
 package asiantech.internship.summer;
 
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.Window;
-import android.view.WindowManager;
-import android.widget.GridView;
+import android.app.Activity;
 
-public class ViewActivity extends AppCompatActivity {
-
-    GridView gridView;
-
-    String[] values = {
-            "A City Living Under The Shawdow",
-            "One Problem for Democratic Leaders",
-            "The Golden Secret to Better Breakfast",
-            "How to Plan Your First Ski Vaction",
-            "How Social Isolation Is Killing Us",
-            "Use Labels to Sort Messages in Facebook"
-    };
-
-
-    int[] images ={
-            R.drawable.avt,
-            R.drawable.avt1,
-            R.drawable.avt2,
-            R.drawable.avt3,
-            R.drawable.avt,
-            R.drawable.avt2
-    };
-    String[] values1 = {
-            "RBC News",
-            "NY Times",
-            "BBC World",
-            "NBC Nightly",
-            "RBC News",
-            "BBC World"
-    };
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        requestWindowFeature(Window.FEATURE_NO_TITLE);
-        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
-        setContentView(R.layout.activity_view);
-
-        gridView = findViewById(R.id.gv1);
-        GridAdapter gridAdapter = new GridAdapter(this, values, images, values1);
-        gridView.setAdapter(gridAdapter);
-    }
+public class ViewActivity extends Activity {
 }
-
