+++ app/src/main/res/layout/activity_main.xml
@@ -5,11 +5,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity">
-    <Button
-        android:id="@+id/btn"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/list"
-        android:background="@color/colorBlue"
-        android:layout_centerInParent="true"/>
 </RelativeLayout>
